+++ simple_history/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import unicode_literals
 
-__version__ = '1.5.3'
+__version__ = '1.5.4'
 
 
 def register(